Eudat ipfs cache is added but mount is not well working.

+++ owncloudScripts/eudatSubmitJob.sh
@@ -7,13 +7,16 @@ if [ "$isOcMounted" != "True" ];then
 fi
 
 ## Share Zip File
-sudo chmod -R 777 exampleFolderToShare/
-sudo tar zcf --SORT=inode exampleFolderToShare.tar.gz exampleFolderToShare
-hash=$(md5sum exampleFolderToShare.tar.gz | awk '{print $1}')
-echo -e 'hash='$hash
-mv exampleFolderToShare.tar.gz $hash.tar.gz
-sudo rsync -avhW --progress $hash.tar.gz $HOME/oc/$hash/
-rm -f $hash.tar.gz
+if [ $# -eq 0 ]; then
+    sudo chmod -R 777 exampleFolderToShare/
+    sudo tar zcf exampleFolderToShare.tar.gz exampleFolderToShare
+    hash=$(md5sum exampleFolderToShare.tar.gz | awk '{print $1}')
+    echo -e 'hash='$hash
+    mv exampleFolderToShare.tar.gz $hash.tar.gz
+    sudo rsync -avhW --progress $hash.tar.gz $HOME/oc/$hash/
+    rm -f $hash.tar.gz
+fi
+
 hash='1c40dae978e1aea987400b145ca48cb9'
 ./singleFolderShare.py $hash
